Four queue implementations.
The simple list-based implementation is actually quite fast, especially
for small queue sizes.

The function- and ST-based implementations were actually a bad idea;
each additional operation increased the size of the function or ST
procedure, causing the implentations to become slower and slower
regardless of the size of the queue.

Finally, the Tree-based log(n) implentation is the fastest for large
queue sizes.